+++ src/main/java/net/lahuta/trdman/repository/TradeRepository.java
@@ -10,7 +10,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 public interface TradeRepository extends JpaRepository<Trade, Long> {
 
     @Query("select i from Trade i where i.id = ?1")
-    Optional<Trade> findById(int id);
+    Optional<Trade> findById(long id);
 
     @Query(nativeQuery = true, value = "select count(*) from trdbook")
     long countItems();
